The US Map is now stored in db. Need to add the states into db.

+++ src/main/java/com/server/server/repository/USMapRepository.java
@@ -3,6 +3,6 @@ package com.server.server.repository;
 import com.server.server.model.map.USMap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface USMapRepository extends JpaRepository<USMap, Long> {
+public interface USMapRepository extends JpaRepository<USMap, Integer> {
 
 }
